#116: Rename slice to layout
